Beta V1.2: Vastly improved 'textWrap.getTextObjects' method to auto wrap passed in text to it. (con't)
This auto wrap is aware of words and very-conservatively does not wrap in the middle of a word. Happy with this code improvement, but lots of code needs cleaning up in the project before the code is a lot more readable. It is readable enough right now, but can definitely be improved a lot.